Moving iree/base/flatcc.h to iree/base/internal/.
