+++ iree/base/CMakeLists.txt
@@ -52,30 +52,6 @@ iree_cc_library(
 
 iree_cc_library(
   NAME
-    flatcc
-  HDRS
-    "flatcc.h"
-  DEPS
-    ::flatcc_dummy
-    flatcc::runtime
-  PUBLIC
-)
-
-flatbuffer_c_library(
-  NAME
-    flatcc_dummy
-  SRCS
-    "flatcc.fbs"
-  FLATCC_ARGS
-    "--reader"
-    "--builder"
-    "--verifier"
-    "--json"
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
     logging
   HDRS
     "logging.h"
