+++ iree/base/internal/CMakeLists.txt
@@ -179,6 +179,30 @@ iree_lit_test_suite(
 
 iree_cc_library(
   NAME
+    flatcc
+  HDRS
+    "flatcc.h"
+  DEPS
+    ::flatcc_dummy
+    flatcc::runtime
+  PUBLIC
+)
+
+flatbuffer_c_library(
+  NAME
+    flatcc_dummy
+  SRCS
+    "flatcc.fbs"
+  FLATCC_ARGS
+    "--reader"
+    "--builder"
+    "--verifier"
+    "--json"
+  PUBLIC
+)
+
+iree_cc_library(
+  NAME
     main
   HDRS
     "main.h"
