+++ iree/base/internal/flatcc.fbs
@@ -18,7 +18,7 @@ namespace iree;
 // when pulled in via generated headers. So here we give ourselves something to
 // include that's always available and cheap.
 //
-// Instead of directly including this file use iree/base/flatcc.h.
+// Instead of directly including this file use iree/base/internal/flatcc.h.
 //
 // Normally including any generated file will include the appropriate headers in
 // the required order (as they are non-hermetic), but that requires that we have
