+++ iree/hal/cuda/BUILD
@@ -65,12 +65,12 @@ cc_library(
         ":dynamic_symbols",
         "//iree/base",
         "//iree/base:core_headers",
-        "//iree/base:flatcc",
         "//iree/base:logging",
         "//iree/base:status",
-        "//iree/base/internal:synchronization",
         "//iree/base:tracing",
         "//iree/base/internal",
+        "//iree/base/internal:flatcc",
+        "//iree/base/internal:synchronization",
         "//iree/hal",
         "//iree/schemas:cuda_executable_def_c_fbs",
     ],
