+++ iree/hal/cuda/CMakeLists.txt
@@ -49,8 +49,8 @@ iree_cc_library(
     ::dynamic_symbols
     iree::base
     iree::base::core_headers
-    iree::base::flatcc
     iree::base::internal
+    iree::base::internal::flatcc
     iree::base::internal::synchronization
     iree::base::logging
     iree::base::status
