+++ iree/vm/CMakeLists.txt
@@ -168,8 +168,8 @@ iree_cc_library(
     ::vm
     iree::base
     iree::base::core_headers
-    iree::base::flatcc
     iree::base::internal
+    iree::base::internal::flatcc
     iree::base::tracing
     iree::schemas::bytecode_module_def_c_fbs
   PUBLIC
